List of data in the EVERSpatDat library

List of data in the EVERSpatDat library

make_EVERSpat()

Value

list of available Everglades related shapefiles.

Examples

make_EVERSpat()
#>       Package      
#>  [1,] "EVERSpatDat"
#>  [2,] "EVERSpatDat"
#>  [3,] "EVERSpatDat"
#>  [4,] "EVERSpatDat"
#>  [5,] "EVERSpatDat"
#>  [6,] "EVERSpatDat"
#>  [7,] "EVERSpatDat"
#>  [8,] "EVERSpatDat"
#>  [9,] "EVERSpatDat"
#> [10,] "EVERSpatDat"
#> [11,] "EVERSpatDat"
#> [12,] "EVERSpatDat"
#> [13,] "EVERSpatDat"
#> [14,] "EVERSpatDat"
#> [15,] "EVERSpatDat"
#> [16,] "EVERSpatDat"
#> [17,] "EVERSpatDat"
#> [18,] "EVERSpatDat"
#> [19,] "EVERSpatDat"
#> [20,] "EVERSpatDat"
#> [21,] "EVERSpatDat"
#> [22,] "EVERSpatDat"
#> [23,] "EVERSpatDat"
#> [24,] "EVERSpatDat"
#> [25,] "EVERSpatDat"
#> [26,] "EVERSpatDat"
#> [27,] "EVERSpatDat"
#>       LibPath                                                                               
#>  [1,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#>  [2,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#>  [3,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#>  [4,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#>  [5,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#>  [6,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#>  [7,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#>  [8,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#>  [9,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [10,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [11,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [12,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [13,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [14,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [15,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [16,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [17,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [18,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [19,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [20,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [21,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [22,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [23,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [24,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [25,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [26,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#> [27,] "C:/Users/PaulJulian/OneDrive - The Everglades Foundation/Documents/R/win-library/4.1"
#>       Item             
#>  [1,] "BBSEERBound"    
#>  [2,] "CEPP"           
#>  [3,] "CSSSSubpops"    
#>  [4,] "EvPA"           
#>  [5,] "FCELTER_sites"  
#>  [6,] "FLAB"           
#>  [7,] "FWCShore_clip"  
#>  [8,] "LOK"            
#>  [9,] "LOKLitt"        
#> [10,] "LOSOM"          
#> [11,] "LOWRP"          
#> [12,] "SFWMD_Projects" 
#> [13,] "SaltExtent"     
#> [14,] "TTBridge"       
#> [15,] "TribalAreas"    
#> [16,] "US41_US1"       
#> [17,] "WCAs"           
#> [18,] "WERPBound"      
#> [19,] "canals"         
#> [20,] "counties_fl"    
#> [21,] "eaa"            
#> [22,] "lakes"          
#> [23,] "lostman_slough" 
#> [24,] "nps_clipped"    
#> [25,] "refuges_clipped"
#> [26,] "sfwmd_bound"    
#> [27,] "sloughs"        
#>       Title                                                                                  
#>  [1,] "Biscayne Bay and Southeastern Everglades Ecosystem Restoration Project Boundary"      
#>  [2,] "Central Everglades Planning"                                                          
#>  [3,] "Cape Sable Seaside Sparrow Subpopulations"                                            
#>  [4,] "Everglades Protection Area Boundary"                                                  
#>  [5,] "FCE LTER long-term monitoring sites"                                                  
#>  [6,] "Florida Bay FATHOM Basins"                                                            
#>  [7,] "Florida Shoreline 1:12000"                                                            
#>  [8,] "Generalized Lake Okeechobee Boundary"                                                 
#>  [9,] "Generalized Lake Okeechobee Littoral Zone Boundary"                                   
#> [10,] "Lake Okeechobee System Operating Manual"                                              
#> [11,] "Lake Okeechobee Watershed Restoration Project"                                        
#> [12,] "South Florida Water Management District Projects"                                     
#> [13,] "Inland Extent of Saltwater (2018)"                                                    
#> [14,] "Tamiami Trail Bridges"                                                                
#> [15,] "Florida Tribal Areas"                                                                 
#> [16,] "Major Roads for FCE LTER"                                                             
#> [17,] "SFWMD Water Conservation Areas"                                                       
#> [18,] "Western Everglades Restoration Project Boundary"                                      
#> [19,] "South Florida Water Management District Canals"                                       
#> [20,] "Florida County Boundaries"                                                            
#> [21,] "SFWMD Everglades Agricultural Area (EAA) Boundary"                                    
#> [22,] "Lakes"                                                                                
#> [23,] "Generalized outline of Lostmans Slough"                                               
#> [24,] "National Park Service areas in Everglades region"                                     
#> [25,] "US Fish and Wildlife Service Refuges"                                                 
#> [26,] "South Florida Water Management District Boundary"                                     
#> [27,] "General delineation of Shark River and Taylor Sloughs within Everglades National Park"
#> attr(,"class")
#> [1] "EVERSpat"